pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secondbrain-mcp"
version = "0.1.0"
description = "MCP Server for SecondBrain with priority-based semantic search"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "fallleaf", email = "fallleaf@example.com"}
]
keywords = ["mcp", "obsidian", "secondbrain", "semantic-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "fastembed>=0.8.0",
    "sqlite-vec>=0.1.7",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "watchdog>=3.0.0",
    "rank-bm25>=0.2.2",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
    "markdown>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
secondbrain-mcp = "secondbrain.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
